+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file. This projec
 
 ## [Unreleased][]
 ### Features
+- **State**: `extensions config` formatting options [#159][]
 - **State**: `extensions run` respects global color/format options [#160][]
 - **State**: delete an extension quietly [#162][]
 
+[#159]: https://github.com/Brickster/git-commands/issues/159
 [#160]: https://github.com/Brickster/git-commands/issues/160
 [#162]: https://github.com/Brickster/git-commands/issues/162